requirements.txt
streamlit>=1.32
openai-whisper
torch
transformers
sentencepiece
gTTS
